Make the offsets pointer sized

+++ filesystem/apps/editor/main.rs
@@ -10,7 +10,7 @@ pub struct Editor {
     url: String,
     file: Option<File>,
     string: String,
-    offset: u32,
+    offset: usize,
     scroll_x: i32,
     scroll_y: i32,
 }
@@ -177,13 +177,13 @@ impl Editor {
                     match key_event.scancode {
                         K_ESC => break,
                         K_BKSP => if self.offset > 0 {
-                            self.string = self.string[0..self.offset as usize - 1].to_string() +
-                                          &self.string[self.offset as usize..self.string.len()];
+                            self.string = self.string[0..self.offset - 1].to_string() +
+                                          &self.string[self.offset..self.string.len()];
                             self.offset -= 1;
                         },
-                        K_DEL => if (self.offset as usize) < self.string.len() {
-                            self.string = self.string[0..self.offset as usize].to_string() +
-                                          &self.string[self.offset as usize + 1..self.string.len() - 1];
+                        K_DEL => if (self.offset) < self.string.len() {
+                            self.string = self.string[0..self.offset].to_string() +
+                                          &self.string[self.offset + 1..self.string.len() - 1];
                         },
                         K_F5 => self.reload(),
                         K_F6 => self.save(&window),
@@ -205,13 +205,13 @@ impl Editor {
                         K_LEFT => if self.offset > 0 {
                             self.offset -= 1;
                         },
-                        K_RIGHT => if (self.offset as usize) < self.string.len() {
+                        K_RIGHT => if (self.offset) < self.string.len() {
                             self.offset += 1;
                         },
-                        K_END => self.offset = self.string.len() as u32,
+                        K_END => self.offset = self.string.len(),
                         K_DOWN => {
                             let mut new_offset = self.string.len();
-                            for i in self.offset as usize..self.string.len() {
+                            for i in self.offset..self.string.len() {
                                 match self.string.as_bytes()[i] {
                                     0 => break,
                                     10 => {
@@ -221,14 +221,14 @@ impl Editor {
                                     _ => (),
                                 }
                             }
-                            self.offset = new_offset as u32;
+                            self.offset = new_offset;
                         }
                         _ => match key_event.character {
                             '\0' => (),
                             _ => {
-                                self.string = self.string[0..self.offset as usize].to_string() +
+                                self.string = self.string[0..self.offset].to_string() +
                                               &key_event.character.to_string() +
-                                              &self.string[self.offset as usize..self.string.len()];
+                                              &self.string[self.offset..self.string.len()];
                                 self.offset += 1;
                             }
                         },
